chore(connect): etract all OpenAPI doc model customisations to a reusable object
