+++ prism-agent/service/server/src/test/scala/io/iohk/atala/api/util/Tapir2StaticOAS.scala
@@ -1,8 +1,9 @@
 package io.iohk.atala.api.util
 
 import io.iohk.atala.agent.server.AgentHttpServer
+import io.iohk.atala.agent.server.http.DocModels
 import io.iohk.atala.castor.controller.{DIDController, DIDRegistrarController}
-import io.iohk.atala.connect.controller.ConnectionController
+import io.iohk.atala.connect.controller.{ConnectionController, ConnectionEndpoints}
 import io.iohk.atala.event.controller.EventController
 import io.iohk.atala.iam.authentication.DefaultAuthenticator
 import io.iohk.atala.iam.entity.http.controller.EntityController
@@ -13,6 +14,7 @@ import io.iohk.atala.pollux.credentialschema.controller.{CredentialSchemaControl
 import io.iohk.atala.presentproof.controller.PresentProofController
 import io.iohk.atala.system.controller.SystemController
 import org.scalatestplus.mockito.MockitoSugar.*
+import sttp.apispec.Tag
 import sttp.tapir.docs.openapi.OpenAPIDocsInterpreter
 import zio.{Scope, ZIO, ZIOAppArgs, ZIOAppDefault, ZLayer}
 
@@ -28,7 +30,9 @@ object Tapir2StaticOAS extends ZIOAppDefault {
       allEndpoints <- AgentHttpServer.agentRESTServiceEndpoints
     } yield {
       import sttp.apispec.openapi.circe.yaml.*
-      val yaml = OpenAPIDocsInterpreter().toOpenAPI(allEndpoints.map(_.endpoint), "Prism Agent", args(1)).toYaml
+      val yaml = DocModels
+        .customiseDocsModel(OpenAPIDocsInterpreter().toOpenAPI(allEndpoints.map(_.endpoint), "Prism Agent", args(1)))
+        .toYaml
       val path = Path.of(args.head)
       Using(Files.newBufferedWriter(path, StandardCharsets.UTF_8)) { writer => writer.write(yaml) }
     }
